# use counters for small & medium ratchet fields