very nice pickup from the old approach of serializing ceiling shasums. encoded ratchets drop from 216 -> 133 bytes. That'll save 8.3Kib over 100 files.

example _old_ serialized ratchet:
```
jiAjzIubJ5xfbrA5OKv5Nd3pO-m_3ABqD1cFNf2oLviZAJ4prGVG6845XUubOkwO6vqLOoFMQ0Ohp6R4k2VeaQRtEJddpyHw6SojKe8S5uU2F6I7q_ZSsWZT6wWXdQdYwYqNp_UdIye5cS9jw1e9Yxbfc33wSTpLGOxpHezlWckjYF35WHDewaLeUePegN0ILY3e07gEHzRmfhzOikjj0Q==
```

example _new_ serialized ratchet:
```
uFwYqNp_UdIye5cS9jw1e9Yxbfc33wSTpLGOxpHezlWckAmQCeKaxlRuvOOV1LmzpMDur6izqBTENDoaekeJNlXmkAjiAjzIubJ5xfbrA5OKv5Nd3pO-m_3ABqD1cFNf2oLvg
```

Also, faster:
```
benchmark                                old ns/op     new ns/op     delta
BenchmarkRatchetAdvance256-4             376107        63957         -82.99%
BenchmarkRatchetDeserializeAdvance-4     1014          878           -13.36%

benchmark                                old allocs     new allocs     delta
BenchmarkRatchetAdvance256-4             4              3              -25.00%
BenchmarkRatchetDeserializeAdvance-4     3              3              +0.00%

benchmark                                old bytes     new bytes     delta
BenchmarkRatchetAdvance256-4             256           176           -31.25%
BenchmarkRatchetDeserializeAdvance-4     560           368           -34.29%
```